requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
sentencepiece>=0.1.99
nltk>=3.8.0
sacrebleu>=2.3.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
datasets>=2.14.0
transformers>=4.30.0

